Moved all files to within project1
